requirements.txt
fastapi
uvicorn
transformers
torch
supabase
python-dotenv
